Extract RedirectedOutput into shared code.
