+++ src/Scripting/CSharpTest/InteractiveSessionTests.cs
@@ -2527,26 +2527,5 @@ System.TypedReference c;
         }
 
         #endregion
-
-        private struct OutputRedirect : IDisposable
-        {
-            private readonly TextWriter _oldOut;
-            private readonly StringWriter _newOut;
-
-            public OutputRedirect(IFormatProvider formatProvider)
-            {
-                _oldOut = Console.Out;
-                _newOut = new StringWriter(formatProvider);
-                Console.SetOut(_newOut);
-            }
-
-            public string Output => _newOut.ToString();
-
-            void IDisposable.Dispose()
-            {
-                Console.SetOut(_oldOut);
-                _newOut.Dispose();
-            }
-        }
     }
 }
